+++ CHANGELOG.md
@@ -1,8 +1,38 @@
-## 5.3.0 (YYYY-MM-DD)
+## 5.4.0 (YYYY-MM-DD)
 
 ### Enhancements
 
 * [ObjectServer] Added support for `SyncSession.addStateChangeListener()`.
+
+
+## 5.3.1 (2018-06-19)
+
+### Bug Fixes
+
+* [ObjectServer] Fixed a bug which could potentially flood Realm Object Server with PING messages.
+* Calling `Realm.deleteAll()` on a Realm file that contains more classes than in the schema throws exception (#5745).
+* `Realm.isEmpty()` returning false in some cases, even if all tables part of the schema are empty (#5745).
+* Fixed rare native crash materializing as `Assertion failed: ref + size <= after_ref with (ref, size, after_ref, ndx, m_free_positions.size())` (#5300).
+
+### Internal
+
+* Upgraded to Realm Core 5.6.2
+* Upgraded to Realm Sync 3.5.6
+* Upgraded to Object Store commit `0bcb9643b8fb14323df697999b79c4a5341a8a21`
+
+
+## 5.3.0 (2018-06-12)
+
+### Enhancements
+
+* [ObjectServer] `Realm.compactRealm(config)` now works on synchronized Realms (#5937).
+* [ObjectServer] `SyncConfiguration.compactOnLaunch()` and `SyncConfiguration.compactOnLaunch(callback)` has been added (#5937).
+* Added `RealmQuery.getRealm()`, `RealmResults.getRealm()`, `RealmList.getRealm()` and `OrderedRealmCollectionSnapshot.getRealm()` (#5997).
+
+### Internal
+
+* Upgraded to Realm Core 5.6.0
+* Upgraded to Realm Sync 3.5.2
 
 
 ## 5.2.0 (2018-06-06)
@@ -21,6 +51,7 @@ This has impacted a number of API's. See below for the details.
 * [ObjectServer] Added `SyncUser.createConfiguration(url)`. Realms created this way are query-based Realms by default.
 * [ObjectServer] Added `SyncUser.getDefaultConfiguration()`.
 * The Realm bytecode transformer now supports incremental builds (#3034).
+* Improved speed and allocations when parsing field descriptions in queries (#5547).
 
 ### Bug Fixes
 
@@ -29,13 +60,6 @@ This has impacted a number of API's. See below for the details.
 ### Internal
 
 * Module mediator classes being generated now produces a stable output enabling better support for incremental builds (#3034).
-
-
-## 5.1.1 (YYYY-MM-DD)
-
-### Enhancements
-
-* Improved speed and allocations when parsing field descriptions in queries (#5547).
 
 
 ## 5.1.0 (2018-04-25)